+++ .github/workflows/main.yml
@@ -182,74 +182,3 @@ jobs:
       - name: Dump Docker logs on failure
         uses: jwalton/gh-docker-logs@v1
         if: failure()
-
-  test_clustering:
-    name: Test C* Clustering (Elixir ${{ matrix.elixir }}, OTP ${{ matrix.otp }}, C* ${{ matrix.cassandra_version }}, Native protocol ${{ matrix.cassandra_native_protocol }})
-    runs-on: ubuntu-20.04
-    strategy:
-      fail-fast: false
-      matrix:
-        include:
-          - otp: "24.2"
-            elixir: "1.13"
-            cassandra_version: "4.1"
-            cassandra_native_protocol: "v5"
-          - otp: "24.2"
-            elixir: "1.13"
-            cassandra_version: "4.1"
-            cassandra_native_protocol: "v4"
-          - otp: "24.2"
-            elixir: "1.13"
-            cassandra_version: "3"
-            cassandra_native_protocol: "v3"
-          - otp: "24.2"
-            elixir: "1.13"
-            cassandra_version: "3"
-            cassandra_native_protocol: "v4"
-          # Oldest supported Elixir/OTP matrix. We support down to Erlang 21, but
-          # rustler_precompiled (which we use for nimble_lz4, which we use for testing)
-          # requires OTP 23+.
-          - otp: "23.3"
-            elixir: "1.11"
-            cassandra_version: "3"
-            cassandra_native_protocol: "v3"
-    env:
-      CASSANDRA_VERSION: ${{ matrix.cassandra_version }}
-      CASSANDRA_NATIVE_PROTOCOL: ${{ matrix.cassandra_native_protocol }}
-
-    steps:
-      - name: Clone the repository
-        uses: actions/checkout@v2
-
-      - name: Start Docker and wait for it to be up
-        run: |
-          docker-compose up --detach --build
-          ./test/docker/health-check-services.sh
-
-      - name: Install OTP and Elixir
-        uses: erlef/setup-beam@v1
-        with:
-          otp-version: ${{matrix.otp}}
-          elixir-version: ${{matrix.elixir}}
-
-      - name: Cache dependencies
-        id: cache-deps
-        uses: actions/cache@v2
-        with:
-          path: |
-            deps
-            _build
-          key: ${{ runner.os }}-otp${{ matrix.otp }}-elixir${{ matrix.elixir }}-${{ hashFiles('**/mix.lock') }}
-
-      - name: Install and compile dependencies
-        if: steps.cache-deps.outputs.cache-hit != 'true'
-        run: |
-          mix deps.get --only test
-          mix deps.compile
-
-      - name: Run clustering end-to-end tests
-        run: mix test.clustering
-
-      - name: Dump Docker logs on failure
-        uses: jwalton/gh-docker-logs@v1
-        if: failure()
